fullscreen/api/element-request-fullscreen-consume-user-activation.html

<!DOCTYPE html>
<title>Element#requestFullscreen() consumes user activation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<div id="log"></div>
<script>
promise_test(async (t) => {
    t.add_cleanup(() => {
        if (document.fullscreenElement) return document.exitFullscreen();
    });
    await test_driver.bless("fullscreen");
    assert_true(navigator.userActivation.isActive, "Activation must be active");
    await document.documentElement.requestFullscreen();
    assert_false(navigator.userActivation.isActive, "Activation is consumed");
});
</script>
